permission

+++ resources/views/product/index.blade.php
@@ -8,8 +8,11 @@
 
 <table class="table table-dark table-hover">
     <div>
-        <a href="{{route('product.create')}}" class="btn btn-success">Add</a>
-        <a href="#modalCreate" data-toggle="modal" class="btn btn-primary">Add (With Modals)</a>
+        @can('permission-owner', Auth::user())
+            <!-- <a href="{{route('product.create')}}" class="btn btn-success">Add</a> -->
+            <a href="#modalCreate" data-toggle="modal" class="btn btn-primary">Add (With Modals)</a>
+        @endcan
+
         <div class="modal fade" id="modalCreate" tabindex="-1" role="basic" aria-hidden="true">
             <div class="modal-dialog">
               <div class="modal-content" >
@@ -108,8 +111,11 @@
                     <a href="{{route('product.edit',$p -> id)}}" class="btn btn-success">Edit</a>
                     
                     <a href="#modalEditA" data-toggle="modal" onclick="getEditForm({{$p->id}})" class="btn btn-success">Edit Product Modal</a>
-                    <a href="#" value="DeleteNoReload" class="btn btn-danger"
-                    onclick="if(confirm('Are you sure to delete {{$p->id}}-{{$p->name}}?')) deleteDataRemoveTR({{$p->id}})">Delete No Reload</a>
+
+                    @can('permission-owner', Auth::user())
+                        <a href="#" value="DeleteNoReload" class="btn btn-danger"
+                        onclick="if(confirm('Are you sure to delete {{$p->id}}-{{$p->name}}?')) deleteDataRemoveTR({{$p->id}})">Delete No Reload</a>
+                    @endcan
 
                     <div class="modal fade" id="modalEditA" tabindex="-1" role="basic" aria-hidden="true">
                         <div class="modal-dialog modal-wide">
@@ -123,10 +129,13 @@
 
                     </div>
                     
+                    
                     <form method="POST" action="{{route('product.destroy',$p->id)}}">
                         @csrf
                         @method('DELETE')
-                        <input type="submit" onclick="return confirm('Are you want to delete this?')" value="Delete" class="btn btn-danger">
+                        @can('permission-owner',Auth::user())
+                            <input type="submit" onclick="return confirm('Are you want to delete this?')" value="Delete" class="btn btn-danger">
+                        @endcan
                     </form>
                 </td>
             </tr>
